Relocate quaternion kinematics and scrub docstrings
Move quaternion_derivative, its inverse, and the raw quaternion product
into quaternion.rs, and the axis basis/rotation helpers into a new
attitude::common module, so each function sits with the representation
it belongs to. Public paths are unchanged via the attitude re-exports.

Restate the public docstrings as behavior, arguments, returns, examples
and a closing reference, moving the Diebel reconciliation into code
comments and into the Learn documentation. Split the attitude
kinematics discussion into its own Learn page, drop the CCSDS
composition note from the attitude representation overview, and register
the new page in the navigation.
