+++ src/attitude/kinematics.rs
@@ -1,220 +1,38 @@
 /*!
-Attitude kinematics: relations between attitude representations and angular
-velocity.
+The `kinematics` module provides relations between Euler-angle rates and
+angular velocity.
 
-All functions use brahe's attitude conventions: passive (coordinate-frame)
-rotations from a frame A to a frame B, scalar-first quaternions, and the
-Hamilton product in which `a * b` applies `a` first (so
-`R(a * b) = R(b) · R(a)` with `R` the world-to-body direction cosine matrix
-of Diebel (2006), eq. 125). Angular velocity is always the angular velocity
-of frame B relative to frame A, expressed in frame B (body frame), in rad/s.
-
-Equations are cited from Diebel (2006), "Representing Attitude: Euler
-Angles, Unit Quaternions, and Rotation Vectors". Diebel's quaternion product
-(his eq. 102) carries a negative vector cross product; it relates to brahe's
-Hamilton product by operand exchange: brahe `a * b` equals Diebel `b · a`.
-Cited equations are restated under this exchange, and the full
-reconciliation is derived once in the attitude representations
-documentation.
-
-Time derivatives of quaternions are not unit quaternions, so this module
-computes quaternion products on raw scalar-first component vectors rather
-than through [`Quaternion`]'s operators, which normalize their results.
+Angular velocity is always the angular velocity of frame B relative to frame A,
+expressed in frame B, in rad/s, where A and B are the source and target frames
+of the attitude. Euler-angle sequences are labelled in brahe's application
+order, the first letter naming the axis of the first-applied rotation.
 */
 
-use nalgebra::{Vector3, Vector4};
+use nalgebra::Vector3;
 
-use crate::attitude::{EulerAngle, Quaternion, RotationMatrix};
-use crate::constants::AngleFormat;
+use crate::attitude::EulerAngle;
+use crate::attitude::common::{axis_rotation, axis_unit};
 use crate::math::SMatrix3;
 use crate::utils::BraheError;
 
-/// Hamilton product on raw scalar-first component vectors.
+/// Splits a brahe `EulerAngleOrder` into the Diebel matrix-order axis digits
+/// `(i, j, k)` and the corresponding Diebel-order angles `(φ_D, θ_D, ψ_D)`.
 ///
-/// Computes brahe's quaternion product `a * b` (Diebel (2006) eqs. 100-104
-/// with operands exchanged, so the vector cross-product term is positive)
-/// without the unit-norm normalization that [`Quaternion`]'s operator
-/// applies. Required for kinematics, where quaternion derivatives are not
-/// unit quaternions.
-///
-/// # Arguments
-/// - `a`: Left operand, scalar-first `[w, x, y, z]`.
-/// - `b`: Right operand, scalar-first `[w, x, y, z]`.
-///
-/// # Returns
-/// Vector4<f64>: The raw (unnormalized) product `a * b`, scalar-first.
-pub(crate) fn raw_quaternion_product(a: Vector4<f64>, b: Vector4<f64>) -> Vector4<f64> {
-    Vector4::new(
-        a[0] * b[0] - a[1] * b[1] - a[2] * b[2] - a[3] * b[3],
-        a[0] * b[1] + b[0] * a[1] + a[2] * b[3] - a[3] * b[2],
-        a[0] * b[2] + b[0] * a[2] + a[3] * b[1] - a[1] * b[3],
-        a[0] * b[3] + b[0] * a[3] + a[1] * b[2] - a[2] * b[1],
-    )
-}
-
-/// Computes the time derivative of an attitude quaternion from the
-/// body-frame angular velocity.
-///
-/// # Source
-///
-/// Diebel (2006), eq. 157: `q̇ = ½ [0; ω']·q`, where `·` is Diebel's
-/// quaternion product and `ω'` is the angular velocity of the body frame
-/// relative to the reference frame, expressed in the body frame.
-///
-/// # Derivation to brahe convention
-///
-/// Brahe's Hamilton product `a * b` equals Diebel's `b · a` (operand
-/// exchange; see the module documentation). Applying the exchange to
-/// eq. 157 gives the implemented form
-///
-/// `q̇ = ½ · (q * [0; ω])`
-///
-/// with the product evaluated on raw components. No sign changes are
-/// involved; the body-frame rate ω right-multiplies under brahe's product
-/// where it left-multiplies under Diebel's.
-///
-/// # Relation to existing representations
-///
-/// This is the derivative-level counterpart of
-/// [`Quaternion::to_rotation_matrix`] (Diebel eq. 125): differentiating
-/// `R(q(t))` and applying the in-tree convention `[ω]× = −Ṙ Rᵀ`
-/// (`src/frames/custom.rs`) for the world-to-body DCM recovers the same ω
-/// this function consumes.
+/// Brahe labels sequences by application order (first letter applied first);
+/// Diebel (2006) eq. 34 labels by left-to-right matrix order, which is the
+/// reverse. The relabeling is `(i, j, k) = (C, B, A)` and
+/// `(φ_D, θ_D, ψ_D) = (ψ, θ, φ)` for brahe order `ABC` with angles `(φ, θ, ψ)`
+/// — the same exchange `EulerAngleOrder::reversed()` applies in the existing
+/// attitude conversions.
 ///
 /// # Arguments
-/// - `q`: Attitude quaternion transforming frame A to frame B.
-/// - `angular_velocity`: Angular velocity of B relative to A, expressed in
-///   B (rad/s).
+///
+/// * `angles`: Euler angles in brahe application order
 ///
 /// # Returns
-/// Vector4<f64>: Scalar-first quaternion derivative `[q̇s, q̇1, q̇2, q̇3]`
-/// (1/s). Not a unit quaternion.
 ///
-/// # Examples
-/// ```
-/// use brahe::attitude::{Quaternion, quaternion_derivative};
-/// use nalgebra::Vector3;
-///
-/// let q = Quaternion::new(1.0, 0.0, 0.0, 0.0);
-/// let q_dot = quaternion_derivative(&q, Vector3::new(0.0, 0.0, 0.1));
-/// assert!((q_dot[3] - 0.05).abs() < 1e-12);
-/// ```
-pub fn quaternion_derivative(q: &Quaternion, angular_velocity: Vector3<f64>) -> Vector4<f64> {
-    let omega_bar = Vector4::new(
-        0.0,
-        angular_velocity[0],
-        angular_velocity[1],
-        angular_velocity[2],
-    );
-    0.5 * raw_quaternion_product(q.to_vector(true), omega_bar)
-}
-
-/// Recovers the body-frame angular velocity from an attitude quaternion and
-/// its time derivative.
-///
-/// # Source
-///
-/// Diebel (2006), eq. 147: `[0; ω'] = 2 q̇·q̄`, where `·` is Diebel's
-/// quaternion product and `q̄` the conjugate.
-///
-/// # Derivation to brahe convention
-///
-/// With brahe's product `a * b` = Diebel `b · a` (see module docs), eq. 147
-/// becomes the implemented form
-///
-/// `[0; ω] = 2 · (q̄ * q̇)`
-///
-/// evaluated on raw components. The scalar component of the product is zero
-/// for exact inputs (up to floating-point error) and is discarded.
-///
-/// # Relation to existing representations
-///
-/// Exact inverse of [`quaternion_derivative`] for unit `q` with `q̇` tangent
-/// to the unit-quaternion manifold (`q · q̇ = 0`). A radial (norm-drift)
-/// component in `q̇` is deliberately projected out by the discarded scalar
-/// term above; this is the desired behavior when `q̇` comes from numerically
-/// differentiating or integrating a quaternion history, where norm drift is
-/// expected. Otherwise equivalent to extracting ω from `[ω]× = −Ṙ Rᵀ`
-/// (`src/frames/custom.rs`) with `R = R(q)` per
-/// [`Quaternion::to_rotation_matrix`].
-///
-/// # Arguments
-/// - `q`: Attitude quaternion transforming frame A to frame B.
-/// - `q_dot`: Scalar-first quaternion derivative (1/s).
-///
-/// # Returns
-/// Vector3<f64>: Angular velocity of B relative to A, expressed in B
-/// (rad/s).
-///
-/// # Examples
-/// ```
-/// use brahe::attitude::{Quaternion, quaternion_derivative, angular_velocity_from_quaternion_derivative};
-/// use nalgebra::Vector3;
-///
-/// let q = Quaternion::new(1.0, 0.0, 0.0, 0.0);
-/// let omega = Vector3::new(0.02, -0.01, 0.3);
-/// let q_dot = quaternion_derivative(&q, omega);
-/// let recovered = angular_velocity_from_quaternion_derivative(&q, q_dot);
-/// assert!((recovered - omega).norm() < 1e-12);
-/// ```
-pub fn angular_velocity_from_quaternion_derivative(
-    q: &Quaternion,
-    q_dot: Vector4<f64>,
-) -> Vector3<f64> {
-    let conjugate = q.conjugate().to_vector(true);
-    let product = raw_quaternion_product(conjugate, q_dot);
-    2.0 * Vector3::new(product[1], product[2], product[3])
-}
-
-/// Basis vector for one axis letter.
-///
-/// # Arguments
-/// - `axis`: Axis digit, `1` for x, `2` for y, and any other value for z.
-///
-/// # Returns
-/// Vector3<f64>: The unit basis vector for the given axis.
-fn axis_unit(axis: u8) -> Vector3<f64> {
-    match axis {
-        1 => Vector3::new(1.0, 0.0, 0.0),
-        2 => Vector3::new(0.0, 1.0, 0.0),
-        _ => Vector3::new(0.0, 0.0, 1.0),
-    }
-}
-
-/// Elementary passive rotation matrix about one axis letter.
-///
-/// # Arguments
-/// - `axis`: Axis digit, `1` for x, `2` for y, and any other value for z.
-/// - `angle`: Rotation angle, radians.
-///
-/// # Returns
-/// SMatrix3: The elementary rotation matrix `Rx`, `Ry`, or `Rz` about the
-/// given axis, evaluated at `angle`.
-fn axis_rotation(axis: u8, angle: f64) -> SMatrix3 {
-    match axis {
-        1 => RotationMatrix::Rx(angle, AngleFormat::Radians).to_matrix(),
-        2 => RotationMatrix::Ry(angle, AngleFormat::Radians).to_matrix(),
-        _ => RotationMatrix::Rz(angle, AngleFormat::Radians).to_matrix(),
-    }
-}
-
-/// Splits a brahe `EulerAngleOrder` into the Diebel matrix-order axis
-/// digits `(i, j, k)` and the corresponding Diebel-order angles
-/// `(φ_D, θ_D, ψ_D)`.
-///
-/// Brahe labels sequences by application order (first letter applied
-/// first); Diebel (2006) eq. 34 labels by left-to-right matrix order, which
-/// is the reverse. The relabeling is `(i, j, k) = (C, B, A)` and
-/// `(φ_D, θ_D, ψ_D) = (ψ, θ, φ)` for brahe order `ABC` with angles
-/// `(φ, θ, ψ)` — the same exchange `EulerAngleOrder::reversed()` applies in
-/// the existing statics (quaternion.rs, rotation_matrix.rs).
-///
-/// # Arguments
-/// - `angles`: Euler angles in brahe application order.
-///
-/// # Returns
-/// ([u8; 3], [f64; 3]): The Diebel matrix-order axis digits `(i, j, k)` and
-/// the corresponding Diebel-order angles `(φ_D, θ_D, ψ_D)`.
+/// * `(digits, angles)`: The Diebel matrix-order axis digits `(i, j, k)` and the
+///   corresponding Diebel-order angles `(φ_D, θ_D, ψ_D)`
 fn diebel_sequence(angles: &EulerAngle) -> ([u8; 3], [f64; 3]) {
     let digits = angles.order as u16; // e.g. ZYX = 321
     let a = (digits / 100) as u8; // first applied (brahe A)
@@ -224,15 +42,16 @@ fn diebel_sequence(angles: &EulerAngle) -> ([u8; 3], [f64; 3]) {
     ([c, b, a], [angles.psi, angles.theta, angles.phi])
 }
 
-/// Builds Diebel's conjugate Euler-angle rates matrix E′ (Diebel (2006)
-/// eq. 38) for the given attitude, in Diebel ordering.
+/// Builds Diebel's conjugate Euler-angle rates matrix E′ (Diebel (2006) eq. 38)
+/// for the given attitude, in Diebel ordering.
 ///
 /// # Arguments
-/// - `angles`: Euler angles in brahe application order.
+///
+/// * `angles`: Euler angles in brahe application order
 ///
 /// # Returns
-/// SMatrix3: The conjugate Euler-angle rates matrix `E′`, in Diebel
-/// ordering.
+///
+/// * `e_prime`: The conjugate Euler-angle rates matrix `E′`, in Diebel ordering
 fn conjugate_rates_matrix(angles: &EulerAngle) -> SMatrix3 {
     let ([i, j, k], [phi_d, theta_d, _psi_d]) = diebel_sequence(angles);
     let ri = axis_rotation(i, phi_d);
@@ -245,38 +64,15 @@ fn conjugate_rates_matrix(angles: &EulerAngle) -> SMatrix3 {
 
 /// Converts Euler-angle rates to the body-frame angular velocity.
 ///
-/// # Source
-///
-/// Diebel (2006), eqs. 38 and 40: `ω' = E′_ijk(u) u̇` with
-/// `E′_ijk(u) = [ê_i, R_i(φ_D)ê_j, R_i(φ_D)R_j(θ_D)ê_k]`, where `(i, j, k)`
-/// are the sequence axes in Diebel's matrix order and `u = (φ_D, θ_D, ψ_D)`
-/// the angles in that order.
-///
-/// # Derivation to brahe convention
-///
-/// Brahe labels sequences by application order; Diebel by matrix order (the
-/// reverse). For brahe order `ABC` with angles `(φ, θ, ψ)` and rates
-/// `(φ̇, θ̇, ψ̇)`, the Diebel sequence is `(i, j, k) = (C, B, A)` with
-/// `u = (ψ, θ, φ)` and `u̇ = (ψ̇, θ̇, φ̇)` — the same relabeling
-/// `EulerAngleOrder::reversed()` applies in the existing statics. The
-/// elementary rotations `R_i` are brahe's `Rx/Ry/Rz`, whose layouts match
-/// Diebel's eqs. 14-16 exactly, so no sign adjustments arise.
-///
-/// # Relation to existing representations
-///
-/// This is the derivative-level counterpart of
-/// `EulerAngle::to_rotation_matrix`: differentiating the same composition
-/// `R = R_A(φ) applied first, then R_B(θ), then R_C(ψ)` that the statics
-/// implement yields exactly `ω' = E′ u̇`. Consistency with
-/// [`quaternion_derivative`] is enforced by test for all 12 sequences.
-///
 /// # Arguments
-/// - `angles`: Euler angles (radians) in brahe application order.
-/// - `rates`: Angle rates `(φ̇, θ̇, ψ̇)` in the same order (rad/s).
+///
+/// * `angles`: Euler angles in brahe application order. Units: (rad)
+/// * `rates`: Angle rates `(φ̇, θ̇, ψ̇)` in the same order as `angles`. Units: (rad/s)
 ///
 /// # Returns
-/// Vector3<f64>: Angular velocity of frame B relative to frame A, expressed
-/// in B (rad/s).
+///
+/// * `angular_velocity`: Angular velocity of frame B relative to frame A,
+///   expressed in B. Units: (rad/s)
 ///
 /// # Examples
 /// ```
@@ -284,15 +80,15 @@ fn conjugate_rates_matrix(angles: &EulerAngle) -> SMatrix3 {
 /// use brahe::AngleFormat;
 /// use nalgebra::Vector3;
 ///
-/// // Identity attitude: ZYX's Diebel matrix order is (X, Y, Z), so E' is the
-/// // identity and the body-frame angular velocity is just the rates
-/// // reordered from brahe's (phi_dot, theta_dot, psi_dot) into Diebel's
-/// // (psi_dot, theta_dot, phi_dot).
 /// let angles = EulerAngle::new(EulerAngleOrder::ZYX, 0.0, 0.0, 0.0, AngleFormat::Radians);
-/// let rates = Vector3::new(0.1, -0.2, 0.3); // (phi_dot, theta_dot, psi_dot)
+/// let rates = Vector3::new(0.1, -0.2, 0.3);
 /// let omega = euler_rates_to_angular_velocity(&angles, rates);
-/// assert!((omega - Vector3::new(rates[2], rates[1], rates[0])).norm() < 1e-12);
+/// assert!((omega - Vector3::new(0.3, -0.2, 0.1)).norm() < 1e-12);
 /// ```
+///
+/// # References:
+///  1. J. Diebel, *Representing Attitude: Euler Angles, Unit Quaternions, and
+///     Rotation Vectors*, 2006. Eqs. 38 and 40.
 pub fn euler_rates_to_angular_velocity(angles: &EulerAngle, rates: Vector3<f64>) -> Vector3<f64> {
     let e_prime = conjugate_rates_matrix(angles);
     // u̇ in Diebel order = (ψ̇, θ̇, φ̇)
@@ -302,39 +98,21 @@ pub fn euler_rates_to_angular_velocity(angles: &EulerAngle, rates: Vector3<f64>)
 
 /// Converts body-frame angular velocity to Euler-angle rates.
 ///
-/// # Source
-///
-/// Inverse of Diebel (2006) eq. 40: `u̇ = E′_ijk(u)⁻¹ ω'`. Diebel gives the
-/// per-sequence closed-form inverses (his §5 per-sequence listings); this
-/// implementation inverts the eq. 38 matrix directly, which is equivalent.
-///
-/// # Derivation to brahe convention
-///
-/// Same relabeling as [`euler_rates_to_angular_velocity`]; the result is
-/// reordered from Diebel `(ψ̇, θ̇, φ̇)` back to brahe `(φ̇, θ̇, ψ̇)`.
-///
-/// # Relation to existing representations
-///
-/// Exact inverse of [`euler_rates_to_angular_velocity`] away from the
-/// sequence's gimbal-lock singularity: `det E′ = ±cos θ` for
-/// distinct-axis (Tait-Bryan) sequences (singular at θ = ±90°) and
-/// `±sin θ` for repeated-axis sequences (singular at θ = 0 or 180°),
-/// matching Diebel §5's singularity statements.
-///
-/// # Singularity policy
-///
-/// This function returns an error when `|det E′| < 1e-6`. The inverse's
-/// conditioning degrades as roughly `2 / |det E′|` near the singularity, so
-/// this threshold bounds the error amplification at roughly `2e6` and
-/// rejects only inputs within roughly `1e-6` rad of the exact singularity.
+/// Exact inverse of [`euler_rates_to_angular_velocity`] away from the sequence's
+/// gimbal-lock singularity. Sequences with three distinct axes are singular at
+/// `θ = ±90°`; sequences that repeat the first axis are singular at `θ = 0°` and
+/// `θ = 180°`. An error is returned within roughly `1e-6` rad of either
+/// condition.
 ///
 /// # Arguments
-/// - `angles`: Euler angles (radians) in brahe application order.
-/// - `angular_velocity`: Body-frame angular velocity (rad/s).
+///
+/// * `angles`: Euler angles in brahe application order. Units: (rad)
+/// * `angular_velocity`: Body-frame angular velocity. Units: (rad/s)
 ///
 /// # Returns
-/// Result<Vector3<f64>, BraheError>: Angle rates `(φ̇, θ̇, ψ̇)` (rad/s), or
-/// `Err(BraheError::NumericalError)` when `|det E′| < 1e-6` (gimbal lock).
+///
+/// * `rates`: Angle rates `(φ̇, θ̇, ψ̇)` in the same order as `angles`, or
+///   `BraheError::NumericalError` at gimbal lock. Units: (rad/s)
 ///
 /// # Examples
 /// ```
@@ -342,18 +120,25 @@ pub fn euler_rates_to_angular_velocity(angles: &EulerAngle, rates: Vector3<f64>)
 /// use brahe::AngleFormat;
 /// use nalgebra::Vector3;
 ///
-/// // Roundtrip: rates -> angular velocity -> rates recovers the input.
 /// let angles = EulerAngle::new(EulerAngleOrder::ZXZ, 0.5, 0.8, -1.2, AngleFormat::Radians);
 /// let rates = Vector3::new(0.02, 0.13, -0.07);
 /// let omega = euler_rates_to_angular_velocity(&angles, rates);
 /// let recovered = angular_velocity_to_euler_rates(&angles, omega).unwrap();
 /// assert!((recovered - rates).norm() < 1e-10);
 /// ```
+///
+/// # References:
+///  1. J. Diebel, *Representing Attitude: Euler Angles, Unit Quaternions, and
+///     Rotation Vectors*, 2006. Eq. 40.
 pub fn angular_velocity_to_euler_rates(
     angles: &EulerAngle,
     angular_velocity: Vector3<f64>,
 ) -> Result<Vector3<f64>, BraheError> {
     let e_prime = conjugate_rates_matrix(angles);
+    // det E′ is ±cos θ for distinct-axis sequences and ±sin θ for repeated-axis
+    // sequences. The inverse's conditioning degrades as roughly 2 / |det E′|, so
+    // the 1e-6 cutoff bounds error amplification at roughly 2e6 and rejects only
+    // inputs within roughly 1e-6 rad of the exact singularity.
     let det = e_prime.determinant();
     if det.abs() < 1e-6 {
         return Err(BraheError::NumericalError(format!(
@@ -377,156 +162,16 @@ mod tests {
     use serial_test::parallel;
 
     use super::*;
-    use crate::attitude::{EulerAngle, EulerAngleOrder, FromAttitude, ToAttitude};
+    use crate::attitude::{
+        EulerAngle, EulerAngleOrder, FromAttitude, Quaternion,
+        angular_velocity_from_quaternion_derivative,
+    };
     use crate::constants::AngleFormat;
 
-    // Analytic single-axis history: rotation about unit axis n at constant rate w
-    // gives q(t) = [cos(w t / 2), sin(w t / 2) * n] with q_dot known in closed form.
-    fn axis_history(n: Vector3<f64>, w: f64, t: f64) -> (Quaternion, Vector4<f64>) {
-        let half = 0.5 * w * t;
-        let q = Quaternion::new(
-            half.cos(),
-            half.sin() * n[0],
-            half.sin() * n[1],
-            half.sin() * n[2],
-        );
-        let q_dot = Vector4::new(
-            -0.5 * w * half.sin(),
-            0.5 * w * half.cos() * n[0],
-            0.5 * w * half.cos() * n[1],
-            0.5 * w * half.cos() * n[2],
-        );
-        (q, q_dot)
-    }
-
-    #[test]
-    #[parallel]
-    fn test_quaternion_derivative_single_axis() {
-        for n in [
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-            Vector3::new(1.0, 2.0, 3.0).normalize(),
-        ] {
-            let w = 0.37;
-            for t in [0.0, 0.4, 1.9, 5.0] {
-                let (q, q_dot_expected) = axis_history(n, w, t);
-                let q_dot = quaternion_derivative(&q, w * n);
-                for i in 0..4 {
-                    assert_abs_diff_eq!(q_dot[i], q_dot_expected[i], epsilon = 1e-12);
-                }
-            }
-        }
-    }
-
-    #[test]
-    #[parallel]
-    fn test_angular_velocity_from_quaternion_derivative_roundtrip() {
-        let q = Quaternion::from_euler_angle(EulerAngle::new(
-            EulerAngleOrder::ZYX,
-            0.3,
-            -0.7,
-            1.1,
-            AngleFormat::Radians,
-        ));
-        let omega = Vector3::new(0.05, -0.02, 0.4);
-        let q_dot = quaternion_derivative(&q, omega);
-        let recovered = angular_velocity_from_quaternion_derivative(&q, q_dot);
-        for i in 0..3 {
-            assert_abs_diff_eq!(recovered[i], omega[i], epsilon = 1e-12);
-        }
-    }
-
-    // Ties the kinematics to brahe's existing statics via the in-tree convention
-    // [omega']x = -R_dot * R^T (src/frames/custom.rs): a body spinning about its own
-    // z-axis after a fixed offset rotation. q(t) = q0 * q_spin(t) (q0 applied first),
-    // R(t) = Rz(w t) * R0, so omega' = w * e_z exactly.
-    #[test]
-    #[parallel]
-    fn test_quaternion_derivative_matches_rotation_matrix_derivative() {
-        let q0 = Quaternion::from_euler_angle(EulerAngle::new(
-            EulerAngleOrder::XYZ,
-            0.2,
-            0.5,
-            -0.3,
-            AngleFormat::Radians,
-        ));
-        let w = 0.9;
-        let omega = Vector3::new(0.0, 0.0, w);
-        let t = 0.8;
-        let dt = 1e-6;
-
-        let q_at = |tau: f64| {
-            let half = 0.5 * w * tau;
-            let spin = Quaternion::new(half.cos(), 0.0, 0.0, half.sin());
-            q0 * spin
-        };
-
-        // Central difference on raw components with hemisphere continuity enforced
-        let qc = q_at(t).to_vector(true);
-        let mut qp = q_at(t + dt).to_vector(true);
-        let mut qm = q_at(t - dt).to_vector(true);
-        if qp.dot(&qc) < 0.0 {
-            qp = -qp;
-        }
-        if qm.dot(&qc) < 0.0 {
-            qm = -qm;
-        }
-        let q_dot_numeric = (qp - qm) / (2.0 * dt);
-
-        let q_dot = quaternion_derivative(&q_at(t), omega);
-        for i in 0..4 {
-            assert_abs_diff_eq!(q_dot[i], q_dot_numeric[i], epsilon = 1e-8);
-        }
-
-        // Cross-check omega against the matrix route used in src/frames/custom.rs
-        let r = |tau: f64| q_at(tau).to_rotation_matrix().to_matrix();
-        let r_dot = (r(t + dt) - r(t - dt)) / (2.0 * dt);
-        let s = -r_dot * r(t).transpose();
-        let omega_matrix = Vector3::new(
-            (s[(2, 1)] - s[(1, 2)]) / 2.0,
-            (s[(0, 2)] - s[(2, 0)]) / 2.0,
-            (s[(1, 0)] - s[(0, 1)]) / 2.0,
-        );
-        let omega_recovered = angular_velocity_from_quaternion_derivative(&q_at(t), q_dot);
-        for i in 0..3 {
-            assert_abs_diff_eq!(omega_matrix[i], omega[i], epsilon = 1e-8);
-            assert_abs_diff_eq!(omega_recovered[i], omega[i], epsilon = 1e-10);
-        }
-    }
-
-    #[test]
-    #[parallel]
-    fn test_quaternion_derivative_sign_covariance() {
-        let q = Quaternion::from_euler_angle(EulerAngle::new(
-            EulerAngleOrder::ZYX,
-            0.3,
-            -0.7,
-            1.1,
-            AngleFormat::Radians,
-        ));
-        let q_vec = q.to_vector(true);
-        let q_neg = Quaternion::new(-q_vec[0], -q_vec[1], -q_vec[2], -q_vec[3]);
-        let omega = Vector3::new(0.05, -0.02, 0.4);
-
-        let q_dot = quaternion_derivative(&q, omega);
-        let q_dot_neg = quaternion_derivative(&q_neg, omega);
-        for i in 0..4 {
-            assert_abs_diff_eq!(q_dot_neg[i], -q_dot[i], epsilon = 1e-12);
-        }
-
-        let q_dot_vec = quaternion_derivative(&q, omega);
-        let omega_from_pos = angular_velocity_from_quaternion_derivative(&q, q_dot_vec);
-        let omega_from_neg = angular_velocity_from_quaternion_derivative(&q_neg, -q_dot_vec);
-        for i in 0..3 {
-            assert_abs_diff_eq!(omega_from_neg[i], omega_from_pos[i], epsilon = 1e-12);
-        }
-    }
-
-    // Classic aerospace 3-2-1 body-rate map (brahe order ZYX: yaw ψ_z first? No —
-    // brahe ZYX applies Z first: phi about Z (yaw), theta about Y (pitch), psi about
-    // X (roll)). Expected from the standard roll-pitch-yaw result with
-    // p,q,r = body rates and brahe's (phi,theta,psi) = (yaw, pitch, roll):
+    // Classic aerospace 3-2-1 body-rate map (brahe order ZYX applies Z first: phi
+    // about Z (yaw), theta about Y (pitch), psi about X (roll)). Expected from the
+    // standard roll-pitch-yaw result with p,q,r = body rates and brahe's
+    // (phi,theta,psi) = (yaw, pitch, roll):
     //   p = psi_dot - phi_dot*sin(theta)
     //   q = theta_dot*cos(psi) + phi_dot*cos(theta)*sin(psi)
     //   r = -theta_dot*sin(psi) + phi_dot*cos(theta)*cos(psi)
