+++ src/attitude/quaternion.rs
@@ -727,12 +727,121 @@ impl From<RotationMatrix> for Quaternion {
     }
 }
 
+/// Computes the Hamilton product of two quaternions given as raw scalar-first
+/// component vectors.
+///
+/// Unlike [`Quaternion`]'s multiplication operator, the result is not
+/// normalized. This is required by the kinematics functions, whose quaternion
+/// derivatives are not unit quaternions.
+///
+/// # Arguments
+///
+/// * `a`: Left operand, scalar-first `[w, x, y, z]`
+/// * `b`: Right operand, scalar-first `[w, x, y, z]`
+///
+/// # Returns
+///
+/// * `product`: The unnormalized product `a * b`, scalar-first
+pub(crate) fn raw_quaternion_product(a: Vector4<f64>, b: Vector4<f64>) -> Vector4<f64> {
+    Vector4::new(
+        a[0] * b[0] - a[1] * b[1] - a[2] * b[2] - a[3] * b[3],
+        a[0] * b[1] + b[0] * a[1] + a[2] * b[3] - a[3] * b[2],
+        a[0] * b[2] + b[0] * a[2] + a[3] * b[1] - a[1] * b[3],
+        a[0] * b[3] + b[0] * a[3] + a[1] * b[2] - a[2] * b[1],
+    )
+}
+
+/// Computes the time derivative of an attitude quaternion from the body-frame
+/// angular velocity.
+///
+/// # Arguments
+///
+/// * `q`: Attitude quaternion transforming frame A to frame B
+/// * `angular_velocity`: Angular velocity of B relative to A, expressed in B. Units: (rad/s)
+///
+/// # Returns
+///
+/// * `q_dot`: Scalar-first quaternion derivative `[q̇s, q̇1, q̇2, q̇3]`. Not a unit
+///   quaternion. Units: (1/s)
+///
+/// # Examples
+/// ```
+/// use brahe::attitude::{Quaternion, quaternion_derivative};
+/// use nalgebra::Vector3;
+///
+/// let q = Quaternion::new(1.0, 0.0, 0.0, 0.0);
+/// let q_dot = quaternion_derivative(&q, Vector3::new(0.0, 0.0, 0.1));
+/// assert!((q_dot[3] - 0.05).abs() < 1e-12);
+/// ```
+///
+/// # References:
+///  1. J. Diebel, *Representing Attitude: Euler Angles, Unit Quaternions, and
+///     Rotation Vectors*, 2006. Eq. 157.
+pub fn quaternion_derivative(q: &Quaternion, angular_velocity: Vector3<f64>) -> Vector4<f64> {
+    let omega_bar = Vector4::new(
+        0.0,
+        angular_velocity[0],
+        angular_velocity[1],
+        angular_velocity[2],
+    );
+    // Diebel's quaternion product is brahe's Hamilton product with the operands
+    // exchanged, so his q̇ = ½ [0; ω]·q is q̇ = ½ (q * [0; ω]) here. The product
+    // runs on raw components because q̇ is not a unit quaternion.
+    0.5 * raw_quaternion_product(q.to_vector(true), omega_bar)
+}
+
+/// Recovers the body-frame angular velocity from an attitude quaternion and its
+/// time derivative.
+///
+/// Exact inverse of [`quaternion_derivative`] for a unit quaternion whose
+/// derivative is tangent to the unit-quaternion manifold. A radial (norm-drift)
+/// component of `q_dot` is projected out, so quaternion derivatives obtained by
+/// numerical differentiation or integration are accepted.
+///
+/// # Arguments
+///
+/// * `q`: Attitude quaternion transforming frame A to frame B
+/// * `q_dot`: Scalar-first quaternion derivative `[q̇s, q̇1, q̇2, q̇3]`. Units: (1/s)
+///
+/// # Returns
+///
+/// * `angular_velocity`: Angular velocity of B relative to A, expressed in B. Units: (rad/s)
+///
+/// # Examples
+/// ```
+/// use brahe::attitude::{Quaternion, quaternion_derivative, angular_velocity_from_quaternion_derivative};
+/// use nalgebra::Vector3;
+///
+/// let q = Quaternion::new(1.0, 0.0, 0.0, 0.0);
+/// let omega = Vector3::new(0.02, -0.01, 0.3);
+/// let q_dot = quaternion_derivative(&q, omega);
+/// let recovered = angular_velocity_from_quaternion_derivative(&q, q_dot);
+/// assert!((recovered - omega).norm() < 1e-12);
+/// ```
+///
+/// # References:
+///  1. J. Diebel, *Representing Attitude: Euler Angles, Unit Quaternions, and
+///     Rotation Vectors*, 2006. Eq. 147.
+pub fn angular_velocity_from_quaternion_derivative(
+    q: &Quaternion,
+    q_dot: Vector4<f64>,
+) -> Vector3<f64> {
+    // Diebel's [0; ω] = 2 q̇·q̄ is 2 (q̄ * q̇) under the same operand exchange. The
+    // scalar component of the product carries only the norm drift of q̇ and is
+    // discarded.
+    let conjugate = q.conjugate().to_vector(true);
+    let product = raw_quaternion_product(conjugate, q_dot);
+    2.0 * Vector3::new(product[1], product[2], product[3])
+}
+
 #[cfg(test)]
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod tests {
     use super::*;
     use approx::assert_abs_diff_eq;
     use serial_test::parallel;
+
+    use crate::constants::AngleFormat;
 
     #[test]
     #[parallel]
@@ -1109,5 +1218,148 @@ mod tests {
         let r = q.to_rotation_matrix();
 
         assert_eq!(Quaternion::from_rotation_matrix(r), q);
+    }
+
+    // Analytic single-axis history: rotation about unit axis n at constant rate w
+    // gives q(t) = [cos(w t / 2), sin(w t / 2) * n] with q_dot known in closed form.
+    fn axis_history(n: Vector3<f64>, w: f64, t: f64) -> (Quaternion, Vector4<f64>) {
+        let half = 0.5 * w * t;
+        let q = Quaternion::new(
+            half.cos(),
+            half.sin() * n[0],
+            half.sin() * n[1],
+            half.sin() * n[2],
+        );
+        let q_dot = Vector4::new(
+            -0.5 * w * half.sin(),
+            0.5 * w * half.cos() * n[0],
+            0.5 * w * half.cos() * n[1],
+            0.5 * w * half.cos() * n[2],
+        );
+        (q, q_dot)
+    }
+
+    #[test]
+    #[parallel]
+    fn test_quaternion_derivative_single_axis() {
+        for n in [
+            Vector3::new(1.0, 0.0, 0.0),
+            Vector3::new(0.0, 1.0, 0.0),
+            Vector3::new(0.0, 0.0, 1.0),
+            Vector3::new(1.0, 2.0, 3.0).normalize(),
+        ] {
+            let w = 0.37;
+            for t in [0.0, 0.4, 1.9, 5.0] {
+                let (q, q_dot_expected) = axis_history(n, w, t);
+                let q_dot = quaternion_derivative(&q, w * n);
+                for i in 0..4 {
+                    assert_abs_diff_eq!(q_dot[i], q_dot_expected[i], epsilon = 1e-12);
+                }
+            }
+        }
+    }
+
+    #[test]
+    #[parallel]
+    fn test_angular_velocity_from_quaternion_derivative_roundtrip() {
+        let q = Quaternion::from_euler_angle(EulerAngle::new(
+            EulerAngleOrder::ZYX,
+            0.3,
+            -0.7,
+            1.1,
+            AngleFormat::Radians,
+        ));
+        let omega = Vector3::new(0.05, -0.02, 0.4);
+        let q_dot = quaternion_derivative(&q, omega);
+        let recovered = angular_velocity_from_quaternion_derivative(&q, q_dot);
+        for i in 0..3 {
+            assert_abs_diff_eq!(recovered[i], omega[i], epsilon = 1e-12);
+        }
+    }
+
+    // Ties the kinematics to brahe's existing statics via the in-tree convention
+    // [omega']x = -R_dot * R^T (src/frames/custom.rs): a body spinning about its own
+    // z-axis after a fixed offset rotation. q(t) = q0 * q_spin(t) (q0 applied first),
+    // R(t) = Rz(w t) * R0, so omega' = w * e_z exactly.
+    #[test]
+    #[parallel]
+    fn test_quaternion_derivative_matches_rotation_matrix_derivative() {
+        let q0 = Quaternion::from_euler_angle(EulerAngle::new(
+            EulerAngleOrder::XYZ,
+            0.2,
+            0.5,
+            -0.3,
+            AngleFormat::Radians,
+        ));
+        let w = 0.9;
+        let omega = Vector3::new(0.0, 0.0, w);
+        let t = 0.8;
+        let dt = 1e-6;
+
+        let q_at = |tau: f64| {
+            let half = 0.5 * w * tau;
+            let spin = Quaternion::new(half.cos(), 0.0, 0.0, half.sin());
+            q0 * spin
+        };
+
+        // Central difference on raw components with hemisphere continuity enforced
+        let qc = q_at(t).to_vector(true);
+        let mut qp = q_at(t + dt).to_vector(true);
+        let mut qm = q_at(t - dt).to_vector(true);
+        if qp.dot(&qc) < 0.0 {
+            qp = -qp;
+        }
+        if qm.dot(&qc) < 0.0 {
+            qm = -qm;
+        }
+        let q_dot_numeric = (qp - qm) / (2.0 * dt);
+
+        let q_dot = quaternion_derivative(&q_at(t), omega);
+        for i in 0..4 {
+            assert_abs_diff_eq!(q_dot[i], q_dot_numeric[i], epsilon = 1e-8);
+        }
+
+        // Cross-check omega against the matrix route used in src/frames/custom.rs
+        let r = |tau: f64| q_at(tau).to_rotation_matrix().to_matrix();
+        let r_dot = (r(t + dt) - r(t - dt)) / (2.0 * dt);
+        let s = -r_dot * r(t).transpose();
+        let omega_matrix = Vector3::new(
+            (s[(2, 1)] - s[(1, 2)]) / 2.0,
+            (s[(0, 2)] - s[(2, 0)]) / 2.0,
+            (s[(1, 0)] - s[(0, 1)]) / 2.0,
+        );
+        let omega_recovered = angular_velocity_from_quaternion_derivative(&q_at(t), q_dot);
+        for i in 0..3 {
+            assert_abs_diff_eq!(omega_matrix[i], omega[i], epsilon = 1e-8);
+            assert_abs_diff_eq!(omega_recovered[i], omega[i], epsilon = 1e-10);
+        }
+    }
+
+    #[test]
+    #[parallel]
+    fn test_quaternion_derivative_sign_covariance() {
+        let q = Quaternion::from_euler_angle(EulerAngle::new(
+            EulerAngleOrder::ZYX,
+            0.3,
+            -0.7,
+            1.1,
+            AngleFormat::Radians,
+        ));
+        let q_vec = q.to_vector(true);
+        let q_neg = Quaternion::new(-q_vec[0], -q_vec[1], -q_vec[2], -q_vec[3]);
+        let omega = Vector3::new(0.05, -0.02, 0.4);
+
+        let q_dot = quaternion_derivative(&q, omega);
+        let q_dot_neg = quaternion_derivative(&q_neg, omega);
+        for i in 0..4 {
+            assert_abs_diff_eq!(q_dot_neg[i], -q_dot[i], epsilon = 1e-12);
+        }
+
+        let q_dot_vec = quaternion_derivative(&q, omega);
+        let omega_from_pos = angular_velocity_from_quaternion_derivative(&q, q_dot_vec);
+        let omega_from_neg = angular_velocity_from_quaternion_derivative(&q_neg, -q_dot_vec);
+        for i in 0..3 {
+            assert_abs_diff_eq!(omega_from_neg[i], omega_from_pos[i], epsilon = 1e-12);
+        }
     }
 }
